Plotting code refactor
